[creditcooperatif] allow multi type connections to work
The code was only dealing with single type connections ([EU]). I added
an additional request that needs to be done when there are several
connection types (on the website the user is asked to choose between
espace particulier or espace pro for example).

The patch also deals with the fact that the domain name is different for
the first two requests of the login step.

Closes: 6521@zendesk