[societegenerale] Password len modified
Password length use to be 6 digit long, but we had one user whose
pasword was 7 digit long. Hence the modification.
The password login/check stays in the browser for now because we'd need
to dig up credentials for the profesional users as well to ensure that
we have the right set of condition upon modifying the login test and do
a migration script.