weboob.browser.nss: force "sql:" prefix on newer versions
NSS seem to have different behaviors on different distros, e.g. Debian
and CentOS, and it also depends on the NSS version, which creates a
whole matrix of cases.
Try to force SQL database use for >=3.35 and default (probably DBM) for
versions <3.35.

NSS might still ask the infamous question:

    Enter Password or Pin for "NSS Certificate DB":

but deleting the old generated *.db files should solve it.