[bp] Changed encoding to UTF-8 for Loan accounts
The encoding to the page seems to have changed and "PRÊT" appeared as
"PRÃªT" : the consequence was that Loans were not typed anymore and were
therefore scraped as regular accounts whereas the Loan page does not
have the same HTML structure.
Removing all the encodings of the module (incluging the subscriptions
part) worked since all pages are now in charset = UTF-8 on the website.

Closes: 27994@Sibi