[boursorama] Fetch card numbers according to the new website
The boursorama website changed and now the former code user to fetch
card numbers is completely obsolete: we must go to the cards page, fetch
the card's hash using card.url, match this hash with another hash on the
page that will be associated with the card's number.
In addition, I removed the method iter_card_ids() that was never called
and constitutes dead code.

Closes: 13401@sibi