[bolden] Remodelled module with one main account and investments
The money borrowed on Bolden should be considered as investments, no
individual accounts, with the available money retrieved as liquidities.
Only the borrowed money that is still ongoing should be considered, the
other investment lines without valuation are expired investments.
With this model, the sum of investment valuation + liquidities actually
fits the main account balance.

Closes: 7716@zendesk, 8244@zendesk