[cragr] Fetch BGPI life insurance balances and correct investments
The balances of BGPI space Life Insurances were missing because on the
BGPI accounts page, some insurance have a balance value =
"INDISPONIBLE". I did an additional request to get the full account
label as well as the account balance.

In addition, the investments were not correctly fetched when several
Life insurances are available on the BGPI space so I did an additional
request to select the correct account before calling iter_investments().

Closes: 8168@zendesk