[creditmutuel] Repair two cards page navigation
When there is additional cards in the cards space, to get the good history of
the card you need to comeback to
the page where the iter account get the cards information. Indeed, for
the same position of card in the two pages the url, headers and
parameters are exactly the same. Thanks creditmutuel website....

Closes: 26890@sibi