[degiro] Fix: balance, original_val, liquidity
Balance : The account balance and the total amount of the investments in
that same account was not the same because of the wrong dictionnary key.

original_valuation : added parse function to handle correclty
investments in foreign currency

liquidity : replaced the investments that are cash in the iter_investments list by
liquidities.