[binck] Corrected details for new Binck website & improved navigation
Some connections were still bugging since the new website arrived,
especially because of the "home page" with market data that can have
several URLs.
From the home page or the view page, I added an on_load() to go directly
to the accounts page.
Some connections are still using the former website and there are no
accounts on the new website architecture so I re-implemented the former
behavior.

Closes: 8489@zendesk