[binck] Get missing ISIN codes from invest detail page
For some investments lists the ISIN codes are missing in the API, but we
can fetch them on the invest detail page.
When the concerned connection will switch to the new interface, I will
check how SRD invests are organized and if we can fetch missing codes
the same way.

Closes: 8489@zendesk