[creditdunordpee] rewrite using s2e, make it work with OTP / device fingerprinting
Credit du Nord PEE has (now) an enhanced security feature,
using a kind of 'fingerprinting' of the browser device which
is used to detect if you changed device ; and will trigger
the sending of a OTP to email or SMS for authorization of
a new device.
If you fill-in the proper OTP you received, it will allow
this new device from now on.

The current creditdunordpee does not handle this feature,
and I was not able to add it easily.

Instead, I found that this bank is using what looks like a
shared platform with other banks - that is already supported
by weboob (module : s2e for the shared code, and modules:
bnppere, capeasi, erehsbc, esalia for the banks using it)

Thus I rewrote the creditdunordpee module to use s2e shared
code and have the authentication working.

In addition to fixing authentication, the use of s2e module
gives us a new capability (CapBankPockets). Morever, it seems
maintained.