[cragr] date_guesser modified
Transactions DO NOT have a year. its only DD/MM. In some cases, if there
aren't enough
transactions in the account, a transaction dating from last year has its
date incorrectly
guessed (with the date_guesser/ to avoid this we had to reduce
data_guesser.date_max_bump
to 2 in order to prevent such problems, but only for savings accounts.It
is not perfect,
hence the assertion to make sure we aren't returning a future date that
is more than 2 days forward.

At date "2018/12/03", the site only shows one transaction with date
"29/12", so this interpreted as "2018/12/29" (future) instead of
"2017/12/29" (past).
If there were more transactions in the year, this wouldn't happen