[boursorama] Corrected liquidities and investments
Boursorama loves to change the xpaths every 2 weeks so we need to adapt.
Now there are several <h3> on top of each <table> so we need to fetch
the first <h3> and check its text().
And the liquidities can be contained in two different tags: h3/h4 or
div/span.

Closes: 8782@zendesk