[cragr] Fix crash due to missing balance
On this quality website the balance of one life insurance was missing
and was not retrievable in any way on the website itself. It caused a
crash because we expected to land on the BGPIPage. We now do not try to
retrieve missing balance if we are not already on a BGPIPage.

closes: 8688@zendesk
closes: 8846@zendesk
closes: 8359@zendesk
closes: 30787@sibi