[societegenerale] fix iter_subscription
Bug due to duplicate id on the page used for iter_subscription.
We use the page where the documents are available to get all the
different subscritpion. When at the end of the year we have 2 documents
for the same subscription, causing the bug. We now skip the annual
document.

Closes: 33740@zendesk