[cragr] Implemented navigation for new Cragr website
Crédit Agricole now uses an API.
This commit enables site switching for the connections of the new
website and keeps the former behavior for the connections that still use
the previous website.
This commit was tested with the backend and we observed no duplicated
account since all the account IDs correctly match the former website
behavior.
All other methods than iter_accounts() return BrowserUnavailable for
now.

Closes: 9101@zendesk