[cragr] Corrected duplicate Loan IDs and iter_cards
In the accounts JSON, some Loans may have identical IDs, so we fetch the
id_element_contrat instead since it is always unique. This value will be
used in the details JSOn to fetch a unique ID for Loans, as well as the
missing account balances.
I also re-wrote iter_cards to fetch all cards (not just cards from the
main_account) and easily sort between immediate and deferred cards.

Closes: 35420@Sibi