[cragr] Implemented history, coming and summaries for Cards
This commit implements iter_history and iter_coming for deferred cards
with the new API website of crédit agricole.
Note that if there are several cards on the same parent account, it is
not possible to figure out which summary belongs to which card so
summaries are only fetched if the card is unique on its parent account.