[netfinca] Correct problems with liquidities and inv codes
Sometimes the self.page.get_liquidity() returns None and we should not
yield "None" instead of an investment.
When return the investment code, we should always check if this code is
a valid ISIN with is_isin_valid().
Some column names may vary between accounts so I corrected the different
possibilities.