[cragr] Add regexp to login to prevent server errors
On the new cragr website, you cannot POST the login+password with a
space at the end of the password, otherwise the website returns a 500
error.
This commit adds a regexp in module.py to raise an error instead of a
browser unavailable when credentials contain something else than digits.

Closes: 9356@zendesk, 35699@sibi