[cragr] Use go_to_account_space() for get_profile and advisor
We are quite often logged out when trying to switch from one space ot
the other. The go_to_accounts_space() already takes this in
consideration and tries to relogin when it happens, so we should use
this method in get_profile and advisor in order to avoid being
disconnected and not handling it.

Closes: 36099@sibi