[cic] Add banqueprive in new_website
We had a silenced bug. The newebsite variable is assigned during the
visit of account page and crash (not the good page excepted). The state keep the
good value of newebsite and during the next synchronization the
iter_accounts can occurs without bug. Here we match the newebsite
with the banqueprive url.

Closes: 36471@sibi