[barclays] Corrected iter_accounts to fetch IBANs
The is_here() of the LoansPage was catching most other accounts pages,
as a results, we never landed on AbstractAccountPage and therefore,
has_iban() always returned False and we were not fetching the IBANs
anymore.
I changed the is_here so that every account fits with the right Page.
From what I observed, only checking accounts have an IBAN so there is no
need to go to every account's detail page during iter_accounts if we do not
scrape anything on it.
This modification significantly reduces the number of requests and the
duration of iter_accounts().