[milleis] Ignore history for twin accounts that are not in Euros
Milleis offers the possibility to have twin accounts, which are actually
one same account with two different currencies (Euros and another currency).
The iter_accounts works
well but on the website, twin accounts transactions are completely
mixed up, and for now we scrape the exact same history for both accounts so
all twin account transactions are duplicated.
Since the currency of each transaciton is not precised, we cannot
pair transactions to one of the twin accounts.
This patch skips iter_history for twin accounts that are not in euros in
order to avoid transaction duplicates, the transactions will now only be
associated with the Euro twin account.

Closes: 9425@zendesk