[cragr] Retry request when main account balance is unavailable
This commit fixes several bugs in iter_accounts, linked to untyped
accounts and unavailable main account balances.
I factorized the try/except when trying to go the an accounts space.
I also added a try.except on the "cards" requests because it often
returns a 400 error that crashes the whole connection.
The count_spaces xpath was corrected to fit specific professional spaces
too.

Closes: 35460@sibi