[cragr] Avoid crash when redirection to Predica for old website
Sometimes during the iter_history on the old Crédit Agricole website, we
get this immediate: "AttributeError: 'PredicaRedirectPage' object has no attribute 'order_transactions'"
This occurs when we try to call order_transactions whereas we are on the
Predica space, which we don't handle yet. Since all regions will switch
to the new website soon we will not build transactions for the Predica
space from the old website.

Closes: 13242@sibi