[netfinca] Corrected account typing, balance and currency
The filling of account attributes was made in a way that did not always
work and was causing errors such as "AttributeError: 'NoneType' object
has no attribute 'group'" on the currency regex.
In addition, not all accounts are Market so we need to Map() the correct
types depending on the account label (most accounts are called COMPTE
TITRES or COMPTE PEA).

Closes: 30957@sibi,  9774@zendesk