[spirica] Only retrieve invests for 20 first transactions & handle logouts
I factorized the transactions investments into a separated method to
retrieve investments only for the first 20 transactions, otherwise the
spirica website is too slow. The rest of iter_history fetches
transactions from other pages without investments.
I also created a method that checks if we are logged in or not.
The current "transactions_page" page is easier to handle if we make it a
class attribute, so it can be stored and modified by all three methods.