[caissedepargne/cenet] rectified transaction type
The transaction types are retrieved according to the `TypeMouvement` key given by the API.
But this key return the same value for card, check and other transactions.
So the `TypeOperationDisplay` key is more precise.
But sometimes it sents the same value for order and transfer despite the label telling the opposite.
So now, the label is looked at first to determine the type, then if nothing is found, the API key is used.

Closes: 9744@zendesk