[fortuneo] Corrected account id and numbers
it seems that there is a new <div> in the account.id xpath so the
RawText() was crashing with the following error:
"TypeError: sequence item 0: expected string or Unicode, NoDefault
found"
I replaced it with a CleanText().replace(), no need to previously
declare a "number" variable, and I also specified the account.number.
I checked that the account.id are the same that previously fetched and
stored in the apishell.

Closes: 38597@sibi