[boursorama] Change the label path to avoid duplicate text
The previous xpath, with the div "label__name" contain two span: one with the short label (hidden in the website) and one with
the long label. Now, we try to get the long one. If it's empty, we take the content of "label__name" to be sure to have a value.