[societegenerale] Fix rdate and type for account card transaction
We used the Raw pattern to set the rdate and the type of a transaction. The problem is that patter raw set the type to "CARD"
(because we need it for card transaction) and not "DEFERRED_CARD". Now, we use raw pattern to set the rdate and we correct
the transaction type in the browser.