[cmes] make invest & pockets more reliable
Before the patch the element were identified by an ID, but this
ID is not the same depending on the subsite or the number
of accounts.

So now the ID is dynamically fetched and not hardcoded anymore.

Also it (can) handles PERCO where the invests are located in a popup, like for the pockets.