[binck] Corrected module to work for all users
During login, some connections end up on a 404 error when reaching the
accounts page, so we must hceck what URLs are available on the HomePage
before going anywhere.

Once we know that it is a unique account connection and that we handle
the right accounts URL, the rest of the website is the same.

Closes: 10843@zendesk, 10844@zendesk, 11067@zendesk, 11068@zendesk,
11190@zendesk, 11176@zendesk, 11207@zendesk