[OAuth2Mixin] Handle generic error for bad url callback
In some cases, the user enrollment failed (server error, bad
credentials, time limit ...), consequently bank api gives us url with error. Here we add a generic method to OAuth2Mixin in order to catch error. In the future we will be able to raise the good error in the specific module.