[weboob/browser/elements] Implement the empty_xpath attribute
The empty_xpath attribute is used to fetch an explicit message on a page
when there is no account, no transaction or no investment for example.
The goal of this attribute is to avoid silent bugs when an item page has
completely changed and the item_xpath does not match any element
anymore: in these cases we don't scrape any element but we don't crash
either ; whereas if empty_xpath is defined at least we will have a
warning that the page may have changed.