[bred] Fix accounts not scrapped
Accounts were filtered using their ID before it was fully formed
resulting in bad filtering.
I moved the filtering after the ID is fully formed and the module now scrapes all
accounts.
I also added the number of the account to account where it's present to
better reflect what is displayed on the website.

closes : 11436@zendesk