[degiro] Repaired iter_history and used float_to_decimal()
The history_history parsing was completely broken. Now, all transaction
have an 'id' key = 0 so we cannot retrieve proper transaction IDs
anymore. The code to get the amount was wrong since there was a (self)
missing for Field('raw'), we were evulating a Regexp() on the Field
class itself.
We do not scrape transaction.amount if it is in a different currency
than the account; in these cases we only scrape the original_amoutn and
the original_currency.