[americanexpress] Fixed transactions parsing when faulty json
iter_history: For certain connections the date 'statement_end_date' can be missing. We
find it elsewhere though, as it corresponds to the end date the period in which the
transaction finds itself.
iter_coming: one of the 2 json scraped by weboob might not exist for
certain connections, we must avoid it.

Closes: 10988@zendesk