[s2e/bnppere] Fetch Investment.performance_history and ISIN codes
The invest form submission enables redirection to a URL from which we
can access the BNP Wealth API.
This API contains the ISIN codes and the performances for BNP invests.
Performances were also implemented for Esalia, the invest details are
usually found on http://sggestion-ede.com or https://www.lyxorfunds.com.