weboob.browser.browser: allow multiple inheritance with AbstractBrowser
When instanciating an AbstractBrowser the 'AbstractBrowser' class was
actually replaced with the class defined in the PARENT attribute.

This behavior did not allow multiple ineritance as the whole
class __basess__ was replaced with the PARENT.

Moreover when you try to inherit an AbstractBrowser from another
AbstractBrowser, only the first AbstractBrowser class was replaced
with the PARENT attribute.

This patch fixes these behaviors by walking over the full parents tree
to replace AbstractBrowser with the current class defined in the
PARENT attributes.