weboob.applications.boobank: declare utf-8 in OFX formatter
USASCII was plain wrong since we used Unicode codepoints without
restricting them to ASCII set.

UTF-8 might not be always right, because at this stage we're
manipulating strings, not bytes, and we don't really know what low-level
encoding is set for (for example) stdout.

UTF-8 is the most likely encoding though, and can't be worse than
USASCII.

Closes: #294