[creditmutuel/cic] Fixed accounts listing for revolving credits
All "Allure"-type loans are now in the list of revolving credits and rightfully
typed as such. Changes had to
be made elsewhere in the module since the way revolving credits are
managed is cumbersome. They have no parent account to be scraped, this
will not make the module fail.

Closes: 12930@zendesk, 44917@sibi, 44918@sibi