[caissedepargne] Corrected investment.diff value
The diff value that we scraped until now was the diff_percent not
divided by 100, whereas the diff is supposed to be a difference in
euros. In the ciassedepargne JSON, there is a key containing the
variation of the valuation in euros so we can actually fill the diff
correctly.