[societegenerale] Added pagination for investments in MarketPage
Next page request is complex to make since it is handled in js on the website.
First need to get the 'idCptSelect' integer param from a drop-down menu
that matches the account.
Then test if there is really a next page to fetch since "Page suivante" link is
always here.
Finally make a request with the right params.

Closes: 12023@zendesk