[sogecartenet] Added transactions on multiple months
By default the website returns only the transactions of the current
month, but we can choose which month we want in the URL of the CSV
document. So now we go get every transactions in the past until we get
empty datas for 3 months in a row (the website doesn't stop us when
there is no more data in the past, so we have to manually stop asking
for datas).

Closes: 13025@zendesk 8640@zendesk