[boursorama] Fix deferred card transactions date
The 'dateVal' key in the CSV with the list of all the deferred transactions is now empty.
The 'obj_date' for deferred cards is now determined with the list of deferred dates available on the website (calendar.ics).

Fixed 'obj_bdate' to get the correct format from the CSV : YYYY-MM-DD (previously it took YYYY-DD-MM)

Removed `replace_dots` for CleanDecimal in CardHistoryPage because
amounts in the CSV already are in the dot format.

Closes: 46551@sibi