[fortuneo] Corrected Life Insurance balance & investment values
The code had to be adapted to handle the fact that Life Insurances have
a '.' for decimal separator on InvestmentPage, not a ',' like other
accounts.
On the investments table, some values have a ',' some others have a '.'
so we need to verify the format in order to call CleanDecimal.French()
or .SI() accordingly.

Closes: 13308@zendesk, 13320@zendesk, 13331@zendesk, 13333@zendesk