[n26] Rework login, Handle OTP and refresh_token
Since n26 added new SCA on its app, their can be only one device paired to an account, other devices will be asked an OTP each time they try to log.

By reversing the app i found that we can use the 'refresh_token', but since it's not used by the app it might be removed.

closes: 13368@zendesk, 13367@zendesk, 13349@zendesk, 13366@zendesk, 13360@zendesk, 13413@zendesk, 13439@zendesk, 13462@zendesk, 13523@zendesk, 13526@zendesk, 13527@zendesk, 13562@zendesk, 13613@zendesk, 13681@zendesk, 13687@zendesk, 13692@zendesk, 13728@zendesk, 13739@zendesk, 13854@zendesk