[societegenerale] check recipient based on id
Until now, when we look for the recipient before making a transfer
we checked on the recipients iban to identify the one.
But the website accepts that a user add several times a recipient
with the same iban.
So we have to check the iban but the id too
to avoid chosing the "wrong" one.