[yomoni] Fix transaction fetching
The transactions were not fetched because the fetching process
relies on a regex that no longer matched all transactions.

The json now contains the character `&nbsp;` between the amount of the
transaction and the currency symbol in most transactions.

The string given by the json is now filtered to convert HTML entities
before fedding it to the regex.