[lcl] Fixed contracts list gathering
The gathering of contracts list was done on another page (ContractsChoicePage) but it
seems to have changed on the website.
After login we now stop on the page `/outil/UAUT/Accueil/preRoutageLogin` to get the list of contracts.
Added a variable to make a required action on the preRoutageLogin page if the page doesn't return any contracts
(which happens if the connection doesn't have multiple contracts).

Closes: 12953@zendesk