[esalia] Fix nav between several investments accounts popups
Investments performances in this module can be fecthed by launching 3
different popups depending on the user's contrats. When wanting the
investments we first collect the urls of each popup then follow each one
and parse the performances from there. When the user has sggestion.com
and lyxorfunds.com invests with first url fetched being lyxorfunds.com,
at the time to follow it, the request would not be accepted and return a
ClientError. If the urls are fetched the other way around, the requests
would work, this is the most cases behavior. In order to make
lyxorfunds.com requests to work everytime we need the referer from
homepage (AccountsPage).