[cragr] Use fill_env instead of overriding LinearDateGuesser
Get rid of NoCopyLinearDateGuesser because it was only created to change
the behaviour of LinearDateGuesser's __deepcopy__ to avoid the copy of
the object when copying *Elements' env.

Now that you can override this behaviour with fill_env this modification
is no longer needed.

See : e1027be50a506baa6fe525a61896e3cfb9c25e6c