tools.backend: allow multiple inheritances with AbstractModule
When instanciating an AbstractModule the 'AbstractModule' class is
actually replaced with the class defined in the PARENT attribute.

But if the parent is also an AbstractModule, the current behavior
does not resolve it.

This patch fixes this behavior by walking over the full parents tree
to replace AbstractModule with the current class defined in the
PARENT attributes.