[cragr] Wrong amount for loans fixed
The 'balance' attribute of a Loan page ('Capital restant dû' on Cragr) was sometimes incorrectly set to the value of 'total_amount' ('Montant Initial').

On their website, a Cragr loan page is made of different tables of five or six columns. Our cragr module dealt with that with a class derived from ListElement.

The ItemElement.parse method was defined with if statements in the following way:

```
if <there exists a table with 5 columns in the web page>:
    <statement 1>
elif <there exists a table with 6 columns in the web page>:
    <statement 2>
```

This behavior is correct when the page contains ONLY tables of 5 columns or when it contains ONLY tables of 6 columns. To make it work with pages containing both, tables of 5 and 6 columns, the if statement was change this way:
```
if <the current item has 5 columns>:
    <statement 1>
elif <the current item has 6 columns>:
    <statement 2>
```