[cragr] handle an other savings history page
When a savings account does not have transactions, a page appears
telling that there is no transactions.
But this page instead of being titled "HISTORIQUE DES OPERATIONS" is
only titled "OPERATIONS".
So this page was detected as an AccountsPage because not matching the
`is_here`.
Causing an issue at the `iter_perimeter_accounts`, because when looking
for the `refreshed_account`, it may not be on the right page.

So now the pages titled "OPERATIONS" are considered
`OtherSavingsHistoryPage`.

Closes: 49916@sibi