[bp] otp and decoupled transfer not implemented
The transfer with otp or decoupled are not supported yet.
For now I raise a AuthMethodNotImplemented so the transfers are not
returned as pending/done.

I know that the website tries to set an otp or decouple because it asks
to the user to chose among several devices.
Those devices are displayed in a pop-up that is called by the JS.
I cannot catch the pop-up but I know that it is there when the
`ConfirmTransferPage` does not have a button and precises to the user
that he has to click the missing button.

I prefer using twice the same URL instead of creating another page
because I don't know yes the behaviour of the website passed the choice
of the device.
So I use whar already exists instead of duplicating or being forced
inherit from an alreading existing page.