[applications/boobank] keep Value for TransferStep and AddRecipientStep
When I needed to use a `ValueBool` when using a `TransferStep`,
I realized that answering 'yes' to the question reacted as planned
but answering 'no' made no change.
No `False` was returned because the `Value` was not kept as a parameter of the
interrupted method.

I don't know why it was made this way (the commit message does not tell
much).
But now, the `Value` is kept no matter its value.