[creditmutuel] reload state after failing adding a recipient
A strange error occured.
During a transfer, an error was raised, coming from a part of the code made to add recipients.
It happened, because a user failed to add it.
Then weboob took back the old session to try to make a transfer.

We need to avoid using a corrupted session.
The `need_clear_storage` browser attribute already existing, I now use it
to be sure that, while raising an error during the addition of a
recipient, the state si cleared out.

Closes: 52319@sibi