[boursorama] Change loan parsing to handle CLIC
Boursorama implemented a payment in 3 months called CLIC, the url was not handled, most of the page is
similar to other loans but most of the Xpaths are not persistent.
For example some Xpaths are there but empty while other lack of accents.

So I added the URL, changed most of the Xpaths to handle all loans, and
since it does not have a rate, I put a default value for this attribute.