[hsbc] fix history issue when multiple credit cards
when two (or more) credit cards are linked to a same parent account, as card numbers are close to each others it can mess'up the fetch of history providing the transactions from the other card.

fixing in a row a side effect of the same issue preventing to revert CB payment record from parent account to card account to perform monthly reset.

fixing also an issue in the history which was missing some transactions:
if history is being called when no new transaction has been recorded
since last credit card payement then history tab will have only last
period and before last period shown. But currently hsbc module assume
that first tab always contains "coming" entries (which are due in the
future) so it is always skipped for fetching history. So past entries of
the last period are not fetched from history and neither form coming as
they are filtered out due to past date.

Add new mapping for transaction types related to Global Transferts (with own foreign accounts).