[bnporc] Get rid of rdates in the future for coming transactions
The modules uses methods from FrenchTransaction to parse the label
and get the rdate from there, the problem is that if it doesn't
find anything the rdate = date, resulting in transactions with
rdate in the future.

The module now sets the rdate to NotAvailable, so that it holds
date data only if a rdate was found in the label.