[vimeo] repair video search and use youtube-dl, drop CapCollection
youtube-dl is great for finding the video url and more up-to-date than
our code so let's use it.

Search hasn't really changed if we don't use the XML API. I didn't
bother to check whether the XML API still worked, that API wasn't
necessary for search and download to work.

I dropped the whole category and CapCollection stuff because it seemed
buggy. No idea if it can be repaired easily, but at least now the module
works for basic stuff. Collection can be repaired later anyway.