[bouygues] handle case where subscription detail are not available
In some case, the subscription details page, where we get the
information for the subscription lab is not available.
We get a 401 error when getting it.

In this case, we can't get the phone numbers associated to the
subscriptions.

We set then the subscription label to the subscription id without the
phone numbers.