[TwoFactorBrowser] TWOFA_DURATION for the state expire date
The STATE_DURATION was used to create the expire date of the state.
But this value, used to know how much time a session should last,
is used to reset the expire date at each new dump_state.

What we want is to set a fixed date from the first login to the next one.

I created the `TWOFA_DURATION` variable to know how much time a session
should hold.
I also create the `get_expire` method to handle the expire date in it,
without changing the dump_state.
This way, I use `twofa_logged_date` to set an expire date that will change at each new login
and not at each `dump_state`.

The cases, that may generate a dirty session and block it for X days,
must be handled out of the weboob session.
It could be taken care of by boobank or equivalent.