[twofactorbrowser] AUTHENTICATION_METHODS set in init
The AUTHENTICATION_MEHTODS is now set in the init.
This way, it exists as a class attribute and it can use function pointer
without been in the middle of the code.
So I check that is is filled before trying to authenticate.

I applied modifications to RedirectBrowser to take the changes.
handle_polling, handle_otp and handle_sms are preserved.