[cmso] New life insurances page
On some connections of BPE (at least), there are no accounts returned
in the savings accounts json page. We can fetch them in an html page on
the Suravenir website. The page already existed but it was not used.

It doesn't have the technical index of the accounts, which is used
elsewhere in the module. Instead we directly have the url to the details
page.

Also, some accounts that are returned correctly in the API are also on the
Suravenir website. We replace them because their info are slightly out-of-sync,
and balance on Suravenir is more coherent (invests are fetched on the same
website and the account balance is equal to the sum of invests; on the API
it isn't).

Closes: 15672@zendesk