Revert "creditmutuel: adjust the mobile message and compute it only when needed."
This reverts commit e820669e79e255d2039009d921f7e584852aaf98.

Upstream commits hadn't been fully taken when developing 2FA fixes, so
the fixes were made on obsolete creditmutuel module files state.
So the 2 trees diverged, making backporting a pain. The whole management
of how 2FA was handled for this module is a world of pain anyway.
I will just revert the shortest conflicting commit, it is the easiest
solution though not the best. Sorry for that.