[caissedepargne] Factorize login and transfer authentication
The new login mechanism is like transfer authentication and use same URLs and pages.
There is 'context' in json response to know if it's a login or transfer,
but this key is not always here when authentication fail and we don't know if it's for login or transfer.

Refactorize all authentication mechanism to have the context through the code.