[creditmutuel] Remove useless _charset_ at login
This argument is not needed and is causing some troubles for password
with some special chars. This was modified to 'utf-8' to allow users
with a `€` in their password to login, but it broke the login for users
with some other special chars (like `à`). If we remove this parameter
from the login form, it works fine for both.

Closes: 57883@sibi 16425@zendesk