[s2e] Check if password is digital before instanciating VK
On the bnppere module, some customers put letters + digits in their passwords, making the VK crash.
However, if we remove these letters and there are exactly 6 digits
remaining, we sometimes manage to log in. We cannot use a regex in module.py
because bnppere has two browsers (Personeo & Visiogo) and the Visiogo
website does not use a VK.